recommendation/requirements.txt
langchain_milvus
pymilvus
pymilvus[model]
python-dotenv
python-multipart
uvicorn
fastapi
psycopg2-binary
fastapi-cli
langchain